rt: ainvs: clear sorry in ArchTcb_AI
Added a bunch of necessary lemmas to clear tc_invs. However, currently
the precondition is is actually false in the case of fault/timeout
handler caps, due to a limitation of is_cnode_or_valid_arch combined
with checked_insert_tcb_invs.

I am not entirely sure how to fix this, but the sorry on
decode_set_space_wf is currently unresolvable with the given
information.