riscv ainvs: tweak valid_uses invariant to solve sorry in ArchAInvsPre
We previously had the user region from 0 to user_vtop, which does not
necessarily include all canonical addresses in the low range. However, even if
users are not able to map anything above user_vtop, they can still access a
virtual address > user_vtop, and our invariants cover this case. (Either the
address will simply not be mapped or it will be a lookup into the kernel part
of the vspace, i.e. a page fault for the user).

This commit introduces canonical_user as the largest canonical address in the
low range of canonical addresses, which is the range reserved for users.